+++ src/_gettsim_personas/de/grundsicherung_für_erwerbsfähige/__init__.py
@@ -1,7 +1,7 @@
 import datetime
 from pathlib import Path
 
-from gettsim_personas.persona_objects import Persona
+from _gettsim_personas.persona_objects import Persona
 
 Couple1Child = Persona(
     path_to_persona_elements=Path(__file__).parent / "_couple_1_child.py",
