--- src/gettsim_personas/de/__init__.py
@@ -1,6 +1,0 @@
-from gettsim_personas.de import (
-    einkommensteuer_sozialabgaben,
-    grundsicherung_für_erwerbsfähige,
-)
-
-__all__ = ["einkommensteuer_sozialabgaben", "grundsicherung_für_erwerbsfähige"]
