+++ src/gettsim_personas/einkommensteuer_sozialabgaben/__init__.py
@@ -1,0 +1,5 @@
+from _gettsim_personas.de.einkommensteuer_sozialabgaben import (
+    Couple1Child,
+)
+
+__all__ = ["Couple1Child"]
