+++ src/gettsim_personas/grundsicherung_für_erwerbsfähige/__init__.py
@@ -1,0 +1,6 @@
+from _gettsim_personas.de.grundsicherung_für_erwerbsfähige import (
+    Couple1Child,
+    Couple1ChildInKarenzzeit,
+)
+
+__all__ = ["Couple1Child", "Couple1ChildInKarenzzeit"]
