+++ tests/infrastructure/test_persona_objects.py
@@ -5,11 +5,11 @@ import numpy as np
 import pytest
 from numpy.testing import assert_array_equal
 
-from gettsim_personas.persona_elements import (
+from _gettsim_personas.persona_elements import (
     persona_description,
     persona_input_element,
 )
-from gettsim_personas.persona_objects import (
+from _gettsim_personas.persona_objects import (
     _fail_if_active_tt_qnames_overlap,
     _fail_if_bruttolohn_m_linspace_grid_is_invalid,
     _fail_if_not_exactly_one_description_is_active,
