+++ tests/personas_for_testing/persona_elements.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING
 
 import numpy as np
 
-from gettsim_personas.persona_elements import (
+from _gettsim_personas.persona_elements import (
     persona_description,
     persona_input_element,
     persona_pid_element,
