+++ tests/personas_for_testing/persona_elements_with_overlapping_dates.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from gettsim_personas.persona_elements import (
+from _gettsim_personas.persona_elements import (
     persona_description,
     persona_input_element,
     persona_pid_element,
